third_party/blink/web_tests/http/tests/cookies/simple-cookies-expired-expected.txt

CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
This test checks that cookies are correctly set using Expires.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Check that setting a simple cookie works.
PASS cookie is 'test=foobar'.
Check setting a cookie that timed out.
PASS cookie is ''.
PASS successfullyParsed is true

TEST COMPLETE
